Make an extra copy when source and destination of copy are aliased
